+++ exporters/otlp/internal/transform/instrumentation.go
@@ -15,7 +15,7 @@
 package transform
 
 import (
-	commonpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/common/v1"
+	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 )